+++ src/main/java/lo23/battleship/online/CommunicationTest.java
@@ -4,15 +4,15 @@ public class CommunicationTest {
 
     public static void main(String[] args) {
 
-        String host = "127.0.0.1";
+        String host = "172.25.35.108";
         int port = 2345;
         ServerTest ts = new ServerTest(host, port);
         ts.open();
         System.out.println("Serveur initialisé.");
 
         //for(int i = 0; i < 5; i++){
-        ClientConnection cConnection = new ClientConnection(host, port);
-        cConnection.start();
+//        ClientConnection cConnection = new ClientConnection(host, port);
+//        cConnection.start();
         //}
     }
 }
